+++ nutype_macros/src/common/parse/mod.rs
@@ -17,7 +17,7 @@ use syn::{
     token::Paren,
 };
 
-use crate::common::models::SpannedDeriveTrait;
+use crate::common::models::{SpannedDeriveTrait, SpannedDeriveUnsafeTrait};
 
 use super::models::{
     ConstFn, CustomFunction, ErrorTypePath, NewUnchecked, TypedCustomFunction, ValueOrExpr,
@@ -74,6 +74,9 @@ pub struct ParseableAttributes<Sanitizer, Validator> {
 
     /// Parsed from `derive(...)` attribute
     pub derive_traits: Vec<SpannedDeriveTrait>,
+
+    /// Parse from `derive_unsafe(...)` attribute
+    pub derive_unsafe_traits: Vec<SpannedDeriveUnsafeTrait>,
 }
 
 enum ValidateAttr<Validator: Parse + Kinded> {
@@ -233,6 +236,7 @@ impl<Sanitizer, Validator> Default for ParseableAttributes<Sanitizer, Validator>
             const_fn: ConstFn::NoConst,
             default: None,
             derive_traits: vec![],
+            derive_unsafe_traits: vec![],
         }
     }
 }
@@ -305,6 +309,33 @@ where
                         let msg = concat!(
                             "To generate ::new_unchecked() function, the feature `new_unchecked` of crate `nutype` needs to be enabled.\n",
                             "But you ought to know: generally, THIS IS A BAD IDEA.\nUse it only when you really need it."
+                        );
+                        return Err(syn::Error::new(ident.span(), msg));
+                    }
+                }
+            } else if ident == "derive_unsafe" {
+                cfg_if! {
+                    if #[cfg(feature = "derive_unsafe")] {
+                        if input.peek(Paren) {
+                            let content;
+                            parenthesized!(content in input);
+                            let items =
+                                content.parse_terminated(SpannedDeriveUnsafeTrait::parse, Token![,])?;
+                            attrs.derive_unsafe_traits = items.into_iter().collect();
+                        } else {
+                            let msg = "`derive_unsafe(...)` must be used with parenthesis.";
+                            return Err(syn::Error::new(ident.span(), msg));
+                        }
+                    } else {
+                        // The feature is not enabled, so we return an error
+                        let msg = concat!(
+                            "To use derive_unsafe() function, the feature `derive_unsafe` of crate `nutype` needs to be enabled.\n\n",
+                            "DID YOU KNOW?\n",
+                            "It's called `derive_unsafe` because it enables to derive any traits that nutype is not aware of.\n",
+                            "So it is developer's responsibility to ensure that the derived traits do not create a loophole to bypass the constraints.\n",
+                            "As the rule of thumb avoid using `derive_unsafe` with traits that:\n",
+                            "- Create a new instance of the type\n",
+                            "- Mutate the value\n\n",
                         );
                         return Err(syn::Error::new(ident.span(), msg));
                     }