+++ nutype_macros/src/integer/generate/traits/mod.rs
@@ -15,7 +15,7 @@ use crate::{
             gen_impl_trait_into, gen_impl_trait_serde_deserialize, gen_impl_trait_serde_serialize,
             gen_impl_trait_try_from, split_into_generatable_traits,
         },
-        models::TypeName,
+        models::{SpannedDeriveUnsafeTrait, TypeName},
     },
     integer::models::{IntegerDeriveTrait, IntegerGuard, IntegerInnerType},
 };
@@ -27,6 +27,7 @@ pub fn gen_traits<T: ToTokens>(
     generics: &Generics,
     inner_type: &IntegerInnerType,
     traits: HashSet<IntegerDeriveTrait>,
+    unsafe_traits: &[SpannedDeriveUnsafeTrait],
     maybe_default_value: Option<syn::Expr>,
     guard: &IntegerGuard<T>,
 ) -> Result<GeneratedTraits, syn::Error> {
@@ -38,6 +39,7 @@ pub fn gen_traits<T: ToTokens>(
     let derive_transparent_traits = quote! {
         #[derive(
             #(#transparent_traits,)*
+            #(#unsafe_traits,)*
         )]
     };