+++ nutype_macros/src/integer/parse.rs
@@ -42,6 +42,7 @@ where
         const_fn,
         default,
         derive_traits,
+        derive_unsafe_traits,
     } = attrs;
     let raw_guard = IntegerRawGuard {
         sanitizers,
@@ -54,6 +55,7 @@ where
         guard,
         default,
         derive_traits,
+        derive_unsafe_traits,
     })
 }